Add the list_categories module
